libs/bot/requirements.txt
groq
chromadb
wikipedia-api
requests
fastapi
uvicorn[standard]
python-multipart
SpeechRecognition
python-dotenv
